todoapp/requirements.txt
fastapi==0.121.0
pydantic==2.12.4
SQLAlchemy==2.0.44
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
alembic
pytest
pytest-asyncio